[Bronze III] Title: 갈래의 색종이 자르기, Time: 0 ms, Memory: 2032 KB -BaekjoonHub

+++ 백준/Bronze/31472. 갈래의 색종이 자르기/갈래의 색종이 자르기.cc
@@ -6,5 +6,5 @@ int main()
 {
     int n;
     cin >> n;
-    cout << (int)4*sqrt(2*n);
+    cout << 4*sqrt(2*n);
 }
